Add react-router and split pages
