+++ frontend/package.json
@@ -15,7 +15,8 @@
   "dependencies": {
     "lucide-react": "^0.515.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.30.1"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^4.2.1",
