+++ frontend/src/pages/Login.jsx
@@ -1,0 +1,103 @@
+import { useState } from 'react'
+import { Button } from '@/components/ui/button.jsx'
+import { Input } from '@/components/ui/input.jsx'
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card.jsx'
+import { Label } from '@/components/ui/label.jsx'
+import { Alert, AlertDescription } from '@/components/ui/alert.jsx'
+import { Send, Loader2, AlertCircle } from 'lucide-react'
+import { authService } from '@/services/api'
+
+export default function Login({ onSuccess }) {
+  const [loginData, setLoginData] = useState({ username: '', password: '' })
+  const [loginLoading, setLoginLoading] = useState(false)
+  const [loginError, setLoginError] = useState('')
+
+  const handleLogin = async (e) => {
+    e.preventDefault()
+    if (!loginData.username || !loginData.password) {
+      setLoginError('Por favor ingresa usuario y contraseña')
+      return
+    }
+
+    setLoginLoading(true)
+    setLoginError('')
+
+    try {
+      const response = await authService.login(loginData)
+      if (response.success) {
+        onSuccess(loginData.username)
+        setLoginError('')
+      }
+    } catch (error) {
+      setLoginError('Error al iniciar sesión. Verifica tus credenciales.')
+    } finally {
+      setLoginLoading(false)
+    }
+  }
+
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-100 flex items-center justify-center p-4">
+      <Card className="w-full max-w-md shadow-xl">
+        <CardHeader className="space-y-1 text-center">
+          <div className="mx-auto mb-4 w-16 h-16 bg-blue-600 rounded-full flex items-center justify-center">
+            <Send className="w-8 h-8 text-white" />
+          </div>
+          <CardTitle className="text-2xl font-bold text-gray-900">PIsky</CardTitle>
+          <CardDescription className="text-gray-600">
+            Gestión de mensajes promocionales
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <form onSubmit={handleLogin} className="space-y-4">
+            {loginError && (
+              <Alert variant="destructive">
+                <AlertCircle className="h-4 w-4" />
+                <AlertDescription>{loginError}</AlertDescription>
+              </Alert>
+            )}
+            <div className="space-y-2">
+              <Label htmlFor="username">Usuario</Label>
+              <Input
+                id="username"
+                type="text"
+                placeholder="Ingresa tu usuario"
+                value={loginData.username}
+                onChange={(e) => setLoginData({ ...loginData, username: e.target.value })}
+                className="h-11"
+                disabled={loginLoading}
+                required
+              />
+            </div>
+            <div className="space-y-2">
+              <Label htmlFor="password">Contraseña</Label>
+              <Input
+                id="password"
+                type="password"
+                placeholder="Ingresa tu contraseña"
+                value={loginData.password}
+                onChange={(e) => setLoginData({ ...loginData, password: e.target.value })}
+                className="h-11"
+                disabled={loginLoading}
+                required
+              />
+            </div>
+            <Button
+              type="submit"
+              className="w-full h-11 bg-blue-600 hover:bg-blue-700"
+              disabled={loginLoading}
+            >
+              {loginLoading ? (
+                <>
+                  <Loader2 className="w-4 h-4 mr-2 animate-spin" />
+                  Iniciando sesión...
+                </>
+              ) : (
+                'Iniciar sesión'
+              )}
+            </Button>
+          </form>
+        </CardContent>
+      </Card>
+    </div>
+  )
+}
